Parse comments on a background thread
